+++ android-sdk-ui/src/main/java/com/appboy/push/AppboyNotificationStyleFactory.java
@@ -9,7 +9,6 @@ import android.graphics.Bitmap;
 import android.graphics.drawable.Icon;
 import android.os.Build;
 import android.os.Bundle;
-import android.view.Gravity;
 import android.view.View;
 import android.widget.RemoteViews;
 
@@ -21,6 +20,7 @@ import androidx.core.app.NotificationCompat;
 
 import com.appboy.Appboy;
 import com.appboy.Constants;
+import com.appboy.IAppboyNavigator;
 import com.appboy.configuration.AppboyConfigurationProvider;
 import com.appboy.enums.AppboyDateFormat;
 import com.appboy.enums.AppboyViewBounds;
@@ -31,10 +31,10 @@ import com.appboy.support.AppboyLogger;
 import com.appboy.support.DateTimeUtils;
 import com.appboy.support.IntentUtils;
 import com.appboy.support.StringUtils;
+import com.appboy.ui.AppboyNavigator;
 import com.appboy.ui.R;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
 public class AppboyNotificationStyleFactory {
   private static final String TAG = AppboyLogger.getAppboyLogTag(AppboyNotificationStyleFactory.class);
@@ -42,34 +42,9 @@ public class AppboyNotificationStyleFactory {
    * BigPictureHeight is set in
    * https://android.googlesource.com/platform/frameworks/base/+/6387d2f6dae27ba6e8481883325adad96d3010f4/core/res/res/layout/notification_template_big_picture.xml.
    */
-  public static final int BIG_PICTURE_STYLE_IMAGE_HEIGHT = 192;
+  private static final int BIG_PICTURE_STYLE_IMAGE_HEIGHT = 192;
   private static final String STORY_SET_GRAVITY = "setGravity";
   private static final String STORY_SET_VISIBILITY = "setVisibility";
-  private static final String START = "start";
-  private static final String CENTER = "center";
-  private static final String END = "end";
-  private static final Integer[] STORY_FULL_VIEW_XML_IDS;
-
-  static {
-    Integer[] idArray = new Integer[6];
-    idArray[0] = R.id.com_appboy_story_text_view;
-    idArray[1] = R.id.com_appboy_story_text_view_container;
-    idArray[2] = R.id.com_appboy_story_text_view_small;
-    idArray[3] = R.id.com_appboy_story_text_view_small_container;
-    idArray[4] = R.id.com_appboy_story_image_view;
-    idArray[5] = R.id.com_appboy_story_relative_layout;
-    STORY_FULL_VIEW_XML_IDS = idArray;
-  }
-
-  private static final Map<String, Integer> GRAVITY_MAP;
-
-  static {
-    Map<String, Integer> stringToGravityInt = new HashMap<>();
-    stringToGravityInt.put(START, Gravity.START);
-    stringToGravityInt.put(CENTER, Gravity.CENTER);
-    stringToGravityInt.put(END, Gravity.END);
-    GRAVITY_MAP = stringToGravityInt;
-  }
 
   /**
    * Sets the style of the notification if supported.
@@ -168,29 +143,48 @@ public class AppboyNotificationStyleFactory {
   }
 
   /**
-   * Returns a DecoratedCustomViewStyle for push story.
-   *
-   * @param context             Current context.
-   * @param notificationExtras  Notification extras as provided by FCM/ADM.
-   * @param notificationBuilder Must be an instance of the v7 builder.
-   * @return a DecoratedCustomViewStyle that describes the appearance of the push story.
+   * @deprecated Please use {@link #getStoryStyle(BrazeNotificationPayload)}
    */
+  @Deprecated
   public static NotificationCompat.DecoratedCustomViewStyle getStoryStyle(Context context,
                                                                           Bundle notificationExtras,
                                                                           Bundle appboyExtras,
                                                                           NotificationCompat.Builder notificationBuilder) {
-    int pageIndex = getPushStoryPageIndex(notificationExtras);
+    BrazeNotificationPayload payload = new BrazeNotificationPayload(context, new AppboyConfigurationProvider(context), notificationExtras);
+    return getStoryStyle(notificationBuilder, payload);
+  }
+
+  /**
+   * Returns a {@link androidx.core.app.NotificationCompat.DecoratedCustomViewStyle} for push story.
+   *
+   * @param context             Current context.
+   * @param notificationExtras  Notification extras for the current page of the push story.
+   * @param notificationBuilder Notification builder.
+   * @return a {@link androidx.core.app.NotificationCompat.DecoratedCustomViewStyle} that describes the appearance of the push story.
+   */
+  public static NotificationCompat.DecoratedCustomViewStyle getStoryStyle(@NonNull NotificationCompat.Builder notificationBuilder,
+                                                                          @NonNull BrazeNotificationPayload payload) {
+    final Context context = payload.getContext();
+    if (context == null) {
+      AppboyLogger.d(TAG, "Push story page cannot render without a context");
+      return null;
+    }
+    final List<BrazeNotificationPayload.PushStoryPage> pushStoryPages = payload.getPushStoryPages();
+    int pageIndex = payload.getPushStoryPageIndex();
+    BrazeNotificationPayload.PushStoryPage pushStoryPage = pushStoryPages.get(pageIndex);
     RemoteViews storyView = new RemoteViews(context.getPackageName(), R.layout.com_appboy_notification_story_one_image);
-    if (!populatePushStoryPage(storyView, context, notificationExtras, appboyExtras, pageIndex)) {
+    if (!populatePushStoryPage(storyView, payload, pushStoryPage)) {
       AppboyLogger.w(TAG, "Push story page was not populated correctly. Not using DecoratedCustomViewStyle.");
       return null;
     }
 
+    final Bundle notificationExtras = payload.getNotificationExtras();
     NotificationCompat.DecoratedCustomViewStyle style = new NotificationCompat.DecoratedCustomViewStyle();
-    int storyPages = getPushStoryPageCount(notificationExtras);
-    PendingIntent previousButtonPendingIntent = createStoryTraversedPendingIntent(context, notificationExtras, (pageIndex - 1 + storyPages) % storyPages);
+    final int numPages = pushStoryPages.size();
+
+    PendingIntent previousButtonPendingIntent = createStoryTraversedPendingIntent(context, notificationExtras, (pageIndex - 1 + numPages) % numPages);
     storyView.setOnClickPendingIntent(R.id.com_appboy_story_button_previous, previousButtonPendingIntent);
-    PendingIntent nextButtonPendingIntent = createStoryTraversedPendingIntent(context, notificationExtras, (pageIndex + 1) % storyPages);
+    PendingIntent nextButtonPendingIntent = createStoryTraversedPendingIntent(context, notificationExtras, (pageIndex + 1) % numPages);
     storyView.setOnClickPendingIntent(R.id.com_appboy_story_button_next, nextButtonPendingIntent);
     notificationBuilder.setCustomBigContentView(storyView);
 
@@ -350,61 +344,30 @@ public class AppboyNotificationStyleFactory {
     }
   }
 
-  private static PendingIntent createStoryPageClickedPendingIntent(Context context, String uriString, String useWebView, String storyPageId, String campaignId) {
+  private static PendingIntent createStoryPageClickedPendingIntent(@NonNull Context context,
+                                                                   @NonNull BrazeNotificationPayload.PushStoryPage pushStoryPage) {
     Intent storyClickedIntent = new Intent(Constants.APPBOY_STORY_CLICKED_ACTION)
         .setClass(context, AppboyNotificationRoutingActivity.class);
     storyClickedIntent
-        .setFlags(storyClickedIntent.getFlags() | Intent.FLAG_ACTIVITY_NO_HISTORY);
-    storyClickedIntent.putExtra(Constants.APPBOY_ACTION_URI_KEY, uriString);
-    storyClickedIntent.putExtra(Constants.APPBOY_ACTION_USE_WEBVIEW_KEY, useWebView);
-    storyClickedIntent.putExtra(Constants.APPBOY_STORY_PAGE_ID, storyPageId);
-    storyClickedIntent.putExtra(Constants.APPBOY_CAMPAIGN_ID, campaignId);
+        .setFlags(storyClickedIntent.getFlags() | AppboyNavigator.getAppboyNavigator().getIntentFlags(IAppboyNavigator.IntentFlagPurpose.NOTIFICATION_PUSH_STORY_PAGE_CLICK));
+    storyClickedIntent.putExtra(Constants.APPBOY_ACTION_URI_KEY, pushStoryPage.getDeeplink());
+    storyClickedIntent.putExtra(Constants.APPBOY_ACTION_USE_WEBVIEW_KEY, pushStoryPage.getUseWebview());
+    storyClickedIntent.putExtra(Constants.APPBOY_STORY_PAGE_ID, pushStoryPage.getStoryPageId());
+    storyClickedIntent.putExtra(Constants.APPBOY_CAMPAIGN_ID, pushStoryPage.getCampaignId());
     return PendingIntent.getActivity(context, IntentUtils.getRequestCode(), storyClickedIntent, 0);
   }
 
   private static PendingIntent createStoryTraversedPendingIntent(Context context, Bundle notificationExtras, int pageIndex) {
-    Intent storyNextClickedIntent = new Intent(Constants.APPBOY_STORY_TRAVERSE_CLICKED_ACTION).setClass(context, AppboyNotificationUtils.getNotificationReceiverClass());
+    Intent storyNextClickedIntent = new Intent(Constants.APPBOY_STORY_TRAVERSE_CLICKED_ACTION)
+        .setClass(context, AppboyNotificationUtils.getNotificationReceiverClass());
     if (notificationExtras != null) {
       notificationExtras.putInt(Constants.APPBOY_STORY_INDEX_KEY, pageIndex);
       storyNextClickedIntent.putExtras(notificationExtras);
     }
-    return PendingIntent.getBroadcast(context, IntentUtils.getRequestCode(), storyNextClickedIntent, PendingIntent.FLAG_ONE_SHOT);
-  }
-
-  /**
-   * Given a notificationExtras with properly formatted image keys (consecutively numbered beginning
-   * at 0), returns the number of images in the push story.
-   *
-   * @param notificationExtras Notification extras as provided by FCM/ADM.
-   * @return The number of images keyed in the given notificationExtras
-   */
-  @VisibleForTesting
-  static int getPushStoryPageCount(Bundle notificationExtras) {
-    int index = 0;
-    while (pushStoryPageExistsForIndex(notificationExtras, index)) {
-      index++;
-    }
-    return index;
-  }
-
-  @VisibleForTesting
-  static boolean pushStoryPageExistsForIndex(Bundle notificationExtras, int index) {
-    return AppboyNotificationActionUtils.getActionFieldAtIndex(index, notificationExtras,
-        Constants.APPBOY_PUSH_STORY_IMAGE_KEY_TEMPLATE, null) != null;
-  }
-
-  /**
-   * Returns the current page index of the push story.
-   *
-   * @param notificationExtras Notification extras as provided by FCM/ADM.
-   * @return The current page index.
-   */
-  @VisibleForTesting
-  static int getPushStoryPageIndex(Bundle notificationExtras) {
-    if (!notificationExtras.containsKey(Constants.APPBOY_STORY_INDEX_KEY)) {
-      return 0;
-    }
-    return notificationExtras.getInt(Constants.APPBOY_STORY_INDEX_KEY);
+    return PendingIntent.getBroadcast(context,
+        IntentUtils.getRequestCode(),
+        storyNextClickedIntent,
+        PendingIntent.FLAG_ONE_SHOT);
   }
 
   /**
@@ -416,61 +379,64 @@ public class AppboyNotificationStyleFactory {
    * @param index              The index of the story page.
    * @return True if the push story page was populated correctly.
    */
-  private static boolean populatePushStoryPage(RemoteViews view, Context context, Bundle notificationExtras, Bundle appboyExtras, int index) {
-    AppboyConfigurationProvider configurationProvider = new AppboyConfigurationProvider(context);
-
-    // Set up title
-    String pageTitle = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-        notificationExtras, Constants.APPBOY_PUSH_STORY_TITLE_KEY_TEMPLATE);
-
-    // If the title is null or blank, the visibility of the container becomes GONE.
-    if (!StringUtils.isNullOrBlank(pageTitle)) {
-      view.setTextViewText(STORY_FULL_VIEW_XML_IDS[0], HtmlUtils.getHtmlSpannedTextIfEnabled(configurationProvider, pageTitle));
-      String titleGravityKey = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-          notificationExtras, Constants.APPBOY_PUSH_STORY_TITLE_JUSTIFICATION_KEY_TEMPLATE, CENTER);
-      int titleGravity = GRAVITY_MAP.get(titleGravityKey);
-      view.setInt(STORY_FULL_VIEW_XML_IDS[1], STORY_SET_GRAVITY, titleGravity);
-    } else {
-      view.setInt(STORY_FULL_VIEW_XML_IDS[1], STORY_SET_VISIBILITY, View.GONE);
+  private static boolean populatePushStoryPage(@NonNull RemoteViews view,
+                                               @NonNull BrazeNotificationPayload payload,
+                                               @NonNull BrazeNotificationPayload.PushStoryPage pushStoryPage) {
+    final Context context = payload.getContext();
+    if (context == null) {
+      AppboyLogger.d(TAG, "Push story page cannot render without a context");
+      return false;
+    }
+    AppboyConfigurationProvider configurationProvider = payload.getAppboyConfigurationProvider();
+    if (configurationProvider == null) {
+      AppboyLogger.d(TAG, "Push story page cannot render without a configuration provider");
+      return false;
     }
 
-    // Set up subtitle
-    String pageSubtitle = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-        notificationExtras, Constants.APPBOY_PUSH_STORY_SUBTITLE_KEY_TEMPLATE);
-
-    //If the subtitle is null or blank, the visibility of the container becomes GONE.
-    if (!StringUtils.isNullOrBlank(pageSubtitle)) {
-      view.setTextViewText(STORY_FULL_VIEW_XML_IDS[2], HtmlUtils.getHtmlSpannedTextIfEnabled(configurationProvider, pageSubtitle));
-      String subtitleGravityKey = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-          notificationExtras, Constants.APPBOY_PUSH_STORY_SUBTITLE_JUSTIFICATION_KEY_TEMPLATE,
-          CENTER);
-      int subtitleGravity = GRAVITY_MAP.get(subtitleGravityKey);
-      view.setInt(STORY_FULL_VIEW_XML_IDS[3], STORY_SET_GRAVITY, subtitleGravity);
-    } else {
-      view.setInt(STORY_FULL_VIEW_XML_IDS[3], STORY_SET_VISIBILITY, View.GONE);
+    final String bitmapUrl = pushStoryPage.getBitmapUrl();
+    if (StringUtils.isNullOrBlank(bitmapUrl)) {
+      AppboyLogger.d(TAG, "Push story page image url invalid");
+      return false;
     }
+    final Bundle notificationExtras = payload.getNotificationExtras();
 
     // Set up bitmap url
-    String bitmapUrl = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-        notificationExtras, Constants.APPBOY_PUSH_STORY_IMAGE_KEY_TEMPLATE);
     Bitmap largeNotificationBitmap = Appboy.getInstance(context).getAppboyImageLoader()
-        .getPushBitmapFromUrl(context, appboyExtras, bitmapUrl, AppboyViewBounds.NOTIFICATION_ONE_IMAGE_STORY);
+        .getPushBitmapFromUrl(context, notificationExtras, bitmapUrl, AppboyViewBounds.NOTIFICATION_ONE_IMAGE_STORY);
     if (largeNotificationBitmap == null) {
       return false;
     }
-    view.setImageViewBitmap(STORY_FULL_VIEW_XML_IDS[4], largeNotificationBitmap);
+    view.setImageViewBitmap(R.id.com_appboy_story_image_view, largeNotificationBitmap);
+
+    // Set up title
+    final String pageTitle = pushStoryPage.getTitle();
+
+    // If the title is null or blank, the visibility of the container becomes GONE.
+    if (!StringUtils.isNullOrBlank(pageTitle)) {
+      final CharSequence pageTitleText = HtmlUtils.getHtmlSpannedTextIfEnabled(configurationProvider, pageTitle);
+      view.setTextViewText(R.id.com_appboy_story_text_view, pageTitleText);
+      int titleGravity = pushStoryPage.getTitleGravity();
+      view.setInt(R.id.com_appboy_story_text_view_container, STORY_SET_GRAVITY, titleGravity);
+    } else {
+      view.setInt(R.id.com_appboy_story_text_view_container, STORY_SET_VISIBILITY, View.GONE);
+    }
+
+    // Set up subtitle
+    final String pageSubtitle = pushStoryPage.getSubtitle();
+
+    // If the subtitle is null or blank, the visibility of the container becomes GONE.
+    if (!StringUtils.isNullOrBlank(pageSubtitle)) {
+      final CharSequence pageSubtitleText = HtmlUtils.getHtmlSpannedTextIfEnabled(configurationProvider, pageSubtitle);
+      view.setTextViewText(R.id.com_appboy_story_text_view_small, pageSubtitleText);
+      int subtitleGravity = pushStoryPage.getSubtitleGravity();
+      view.setInt(R.id.com_appboy_story_text_view_small_container, STORY_SET_GRAVITY, subtitleGravity);
+    } else {
+      view.setInt(R.id.com_appboy_story_text_view_small_container, STORY_SET_VISIBILITY, View.GONE);
+    }
 
     // Set up story clicked intent
-    String campaignId = notificationExtras.getString(Constants.APPBOY_PUSH_CAMPAIGN_ID_KEY);
-    String storyPageId = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-        notificationExtras, Constants.APPBOY_PUSH_STORY_ID_KEY_TEMPLATE, "");
-    String deepLink = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-        notificationExtras, Constants.APPBOY_PUSH_STORY_DEEP_LINK_KEY_TEMPLATE);
-    String useWebView = AppboyNotificationActionUtils.getActionFieldAtIndex(index,
-        notificationExtras, Constants.APPBOY_PUSH_STORY_USE_WEBVIEW_KEY_TEMPLATE);
-    PendingIntent storyClickedPendingIntent = createStoryPageClickedPendingIntent(context,
-        deepLink, useWebView, storyPageId, campaignId);
-    view.setOnClickPendingIntent(STORY_FULL_VIEW_XML_IDS[5], storyClickedPendingIntent);
+    PendingIntent storyClickedPendingIntent = createStoryPageClickedPendingIntent(context, pushStoryPage);
+    view.setOnClickPendingIntent(R.id.com_appboy_story_relative_layout, storyClickedPendingIntent);
     return true;
   }